requirements.txt
# Jax
# Distributed only working on 0.3.10
--find-links https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
jax[cuda]==0.3.10

# From deepmind-research/adversarial_robustness
# We comment out jax because 0.2.25 or later is required for jax.distributed.initialize
# We comment out torch dependencies because we are using jax.
absl-py==0.12.0
chex==0.0.7
dm-haiku==0.0.4
einops==0.3.0
# jax==0.2.16
# jaxlib==0.1.68
jaxline==0.0.3
ml-collections==0.1.0
numpy==1.19.5
optax==0.0.8
tensorflow==2.5.0
tensorflow-datasets==4.3.0
# torch==1.9.0
# torchvision==0.10.0
tqdm==4.61.1

# Utilities
black==22.3.0
gpustat
